allow loading of >32kb data

+++ game/disk.c
@@ -5,7 +5,7 @@ LoadMFMB(__REG("a0", void* dest), __REG("d0", uint32_t startSector), __REG("d1",
 extern uint32_t startCode;
 
 __EXTERNAL void
-disk_loadData(__REG("a0", void* dest), __REG("a1", void* src), __REG("d0", int16_t size))
+disk_loadData(__REG("a0", void* dest), __REG("a1", void* src), __REG("d0", int32_t size))
 {
 #if TRACKLOADER==1
   int startSector = ((((uint32_t)src)-((uint32_t)&startCode))>>9)+2; // +2 for bootblock
