+++ game/disk.h
@@ -2,6 +2,6 @@
 #define DISK_H
 
 void
-disk_loadData(__REG("a0", void* dest), __REG("a1", void* src), __REG("d0", int16_t size));
+disk_loadData(__REG("a0", void* dest), __REG("a1", void* src), __REG("d0", int32_t size));
 
 #endif
